Base version (not working) of script to create ethnical boundaries.
Currently the script assumes each town is in only one constituency
wheras in fact, we think the constituencies overlap, meaning that
the same town can be in more than one constituency.

We think the constituencies can be created from the towns (rather
than going down to the village level) if this amendment is made.