boundaries/index.json entry added for county-regional-constituencies
This is currently incorrect. the "area_type_wikidata_item_id" value contains
the wikidata id of the entity  which all specific county council constituencies
are subclasses of. The association contains one example of a councillor from
nantou county.

Our g0v partners have only recently created the classes like 'constituency of
nantou county council' so it's not yet clear that it is modelled exactly as we
expect. This query (http://tinyurl.com/y9fed96l)  shows what is probably the
bulk of the detail needed to complete boundaries/index.json, but it should be
reviewed by a wikidata wrangler first if possible.