Add script and json to repository.
the file 'county-constituencies-to-towns.json contains a list of each
constituency grouped by county. Each constituency contains a list of towns
which make up the geographical area. It was generated by g0v partners and then
updated with two ommissions by hand. The towns correspond to those contained
in the source shapefile of administrative districts.
The script make_regional_constituencies.py uses the json mapping to create
polygons representing each of the constituencies. It alters both the attribute
data and the geometries - dissolving the towns to form constituencies.