Add shp and csv for national boundaries.
Currently includes geoms which are unassigned to constituencies.
These are geometries separated by county.

MS_FB and MS_FB_PARE codes have been added with the assumption made that
the constituencies are children of the counties. This is because the JSON
from which the counstituencies were constructed also had them sorted by county.

It seems simpler to undo it if it's not wanted than to have to add them.
Likewise for the attribute data regarding counties and the features unassigned
to a constituency.