Fix up the Hsinchu City constituencies.
These had been divided along village lines rather than districts, so the
original script failed to build them correctly. A new mapping was provided
which detailed the village makeup. This was used to reconstruct them.

There are two of the 'not compiled' villages in the city. These have been
captured everywhere else in the country for this type of constituency.
Since they were not on an internal district divide the decision was made
to include them with the consituency that matched their districts. The
alternative would have made them 'islands' disconnected from their
constituency.